Adding: Exception handling in department entity

+++ week2/department/src/main/java/com/example/department/services/DepartmentService.java
@@ -2,12 +2,13 @@ package com.example.department.services;
 
 import com.example.department.dtos.DepartmentDTO;
 import com.example.department.entities.DepartmentEntity;
+import com.example.department.exceptions.ResourceNotFoundException;
 import com.example.department.repositories.DepartmentRepository;
 import org.modelmapper.ModelMapper;
 import org.springframework.stereotype.Service;
 
-import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service
@@ -21,6 +22,11 @@ public class DepartmentService {
         this.modelMapper = new ModelMapper();
     }
 
+    public Boolean isExist(Long id){
+        Optional<DepartmentEntity> department = departmentRepository.findById(id);
+        return department.isPresent();
+    }
+
     public List<DepartmentDTO> getAllDepartments(){
         return departmentRepository.findAll()
                 .stream()
@@ -29,8 +35,11 @@ public class DepartmentService {
     }
 
     public DepartmentDTO getDepartmentById(Long id){
-        DepartmentEntity department = departmentRepository.findById(id)
-                .orElse(null);
+        Boolean flag = isExist(id);
+        if(!flag)
+            throw new ResourceNotFoundException(" with id" + id);
+
+        Optional<DepartmentEntity> department = departmentRepository.findById(id);
         return modelMapper.map(department, DepartmentDTO.class);
     }
 
@@ -42,13 +51,18 @@ public class DepartmentService {
 
     public DepartmentDTO updateDepartment(Long id, DepartmentDTO departmentDTO){
         DepartmentEntity toUpdate = departmentRepository.findById(id)
-                .orElse(null);
-        modelMapper.map(departmentDTO, toUpdate);
-        DepartmentEntity saved = departmentRepository.save(toUpdate);
-        return modelMapper.map(saved, DepartmentDTO.class);
+                .orElseThrow(() -> new ResourceNotFoundException(" with id " + id));
+
+        toUpdate.setTitle(departmentDTO.getTitle());
+        toUpdate.setIsActive(departmentDTO.getIsActive());
+        toUpdate.setCreationAt(departmentDTO.getCreationAt());
+        return modelMapper.map(departmentRepository.save(toUpdate), DepartmentDTO.class);
     }
 
     public void deleteDepartment(Long id){
-        departmentRepository.deleteById(id);
+       Boolean flag = isExist(id);
+       if(!flag)
+           throw new ResourceNotFoundException("with id " + id);
+       departmentRepository.deleteById(id);
     }
 }
